move config from internal to pkg for consumption downstream
