+++ service/pkg/server/services.go
@@ -13,10 +13,10 @@ import (
 	"github.com/opentdf/platform/service/authorization"
 	"github.com/opentdf/platform/service/entityresolution"
 	"github.com/opentdf/platform/service/health"
-	"github.com/opentdf/platform/service/internal/config"
 	"github.com/opentdf/platform/service/internal/server"
 	"github.com/opentdf/platform/service/kas"
 	logging "github.com/opentdf/platform/service/logger"
+	"github.com/opentdf/platform/service/pkg/config"
 	"github.com/opentdf/platform/service/pkg/db"
 	"github.com/opentdf/platform/service/pkg/serviceregistry"
 	"github.com/opentdf/platform/service/policy"
